rewrite-flatpak-manifest: Only rewrite first source
The script currently rewrites all git sources of the main module
to use dir/path instead. This breaks for secondary sources, for
example subprojects that are placed into the main source tree
to satisfy a meson-wrap dependency.

It should be safe to assume that if a module has multiple sources,
the main source will be first, and following sources modify or
complement those sources.

So only rewrite the first source if it is of type git, and keep
other sources unmodified regardless of their type.

Closes: https://gitlab.gnome.org/GNOME/gnome-runtime-images/-/issues/14